+++ src/main/webapp/WEB-INF/static/js/main.js
@@ -1,0 +1,5 @@
+function edit(id, description, status) {
+    document.getElementById("taskId").value = id;
+    document.getElementById("description").value = description;
+    document.getElementById("status").value = status;
+}
